fix(packages): show CoStakingBoostModal in Multistaking instead of StakingExpansion
